+++ src/eservice/eventsV2.ts
@@ -38,6 +38,8 @@ import {
   EServiceDescriptorDocumentDeletedByTemplateUpdateV2,
   EServiceNameUpdatedByTemplateUpdateV2,
   EServiceDescriptorAgreementApprovalPolicyUpdatedV2,
+  EServiceSignalHubEnabledV2,
+  EServiceSignalHubDisabledV2,
 } from "../gen/v2/eservice/events.js";
 
 export function eServiceEventToBinaryDataV2(
@@ -163,6 +165,12 @@ export function eServiceEventToBinaryDataV2(
       { type: "EServiceDescriptorAgreementApprovalPolicyUpdated" },
       ({ data }) =>
         EServiceDescriptorAgreementApprovalPolicyUpdatedV2.toBinary(data)
+    )
+    .with({ type: "EServiceSignalHubEnabled" }, ({ data }) =>
+      EServiceSignalHubEnabledV2.toBinary(data)
+    )
+    .with({ type: "EServiceSignalHubDisabled" }, ({ data }) =>
+      EServiceSignalHubDisabledV2.toBinary(data)
     )
     .exhaustive();
 }
@@ -454,6 +462,22 @@ export const EServiceEventV2 = z.discriminatedUnion("type", [
     event_version: z.literal(2),
     type: z.literal("EServiceDescriptorAgreementApprovalPolicyUpdated"),
     data: protobufDecoder(EServiceDescriptorAgreementApprovalPolicyUpdatedV2),
+    stream_id: z.string(),
+    version: z.number(),
+    timestamp: z.coerce.date(),
+  }),
+  z.object({
+    event_version: z.literal(2),
+    type: z.literal("EServiceSignalHubEnabled"),
+    data: protobufDecoder(EServiceSignalHubEnabledV2),
+    stream_id: z.string(),
+    version: z.number(),
+    timestamp: z.coerce.date(),
+  }),
+  z.object({
+    event_version: z.literal(2),
+    type: z.literal("EServiceSignalHubDisabled"),
+    data: protobufDecoder(EServiceSignalHubDisabledV2),
     stream_id: z.string(),
     version: z.number(),
     timestamp: z.coerce.date(),