+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 
 All notable changes to this project will be documented in this file.
+
+## 1.6.4
+
+### Added
+
+- Added `EServiceSignalHubEnabled` and `EServiceSignalHubDisabled` eservice event
+
 ## 1.6.3
 
 ### Fixed